bug fix and memory print statements;

+++ src/TrafficEnv.java
@@ -33,7 +33,7 @@ public class TrafficEnv implements Constant {
 	public TrafficEnv(int max_value, int discretization) {
 		this.max_value = max_value;
 		this.discretization = discretization;
-		ratio = max_value / discretization;
+		ratio = (float) max_value / discretization;
 		// this.cols = cols;
 		// stateSizeChange();
 		this.modifyStateSize();
